Split args.hpp into args.hpp and utils.hpp
